fix getLatestOutput of matrix_gui
until now one could often not open the preview when in the output line a
filename with the extension ".ma7" was specified. This should now be
fixed. but I wonder if upon entering the matrix and generation of a
filename one should not change what is written in the output filename
field? @phowl any thoughts?