make control_dummy a gui_script (#136)
* make control_dummy a pure GUI w/o console output

* tested/fixed control_dummy on windows

* pep8formatting action fixes (#137)

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>

* add error handling to the two key parts of the gui script, can still be improved?

* set logging stream only when available

* implement changes according to discussion in #133

* correctly implement refreshDict and errorhandling

One should never touch the GUI elements from threads other than the main
thread. pyqtSignal has to be used to signal any change which has to be
shown in the GUI. The tools for this where already there but not used
correctly. This should be now find in control_dummy, but should be also
changed in the other control_guis. This can cause strange segmentation
faults which appear only once also the main thread accesses the GUI
elements. (hard to trace)

* pep8formatting action fixes (#141)

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>

* remove test exception

tested on windows and it seems to work also on this platform.

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>
Co-authored-by: pheowl <github@rschlitz.de>