Scpi pickling in ASCII compatible way (#143)
* allow binary data to pass through the scpi server

so far one needed to convert them to an utf-8 string which is a
dangerous thing since not all byte values need to safely convert to a
string. At the blue15 system for one particular set of PID parameters
for the Lakeshore we hit such a problem which gave an almost impossible
to debug error.

this still needs testing

* fixes after testing on blue15

the pickled data have to be converted to ascii since pyvisa would
otherwise need be configured to read some bytes data which I could not
see how it works.

For this reason the pickle protocol version 0 has to be used since it
otherwise can often not be converted to a string.

This solution works but IMHO does not represent the ideal solution. If
there would be an easy way to have pymeasure/pyvisa read the binary data
(without converting them to something with numpy) it would be better.

* pep8formatting action fixes (#144)

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>