use exclusively Pyproject.toml and switch to flit build system (#237)
* modernize build system of core_library/matr1x

switching from setuptools to flit which supports PEP660 which we likley
need to rely on in the future (editable installs using a pyproject.toml)

this needs a lot of testing before merging, but seems promissing and it
might address issue #234

* modernize build system for ifwlib

switch to pyproject toml and flit build system to suppport PEP660

* remove calls of setup.py from install script

* remove setup.py/setuptools from the README

* update install scripts

harmonize the printed text. try to have recommended answers as default
answer. I only left the last three answers with a default No anders
since those typically only upon a first install want to be answered by
yes, but the script's use will be more often for repeating an
installation.

* require pip >=21.3 for the build to support PEP660

I am not sure if this works like this since pip itself would need to
update itself and restart with the new version. It needs to be tested.

* run package tests on all supported python versions on github

* update README to add pip requirement

* add requirements text to update pip before INSTALL

* simplify windows installation instructions and script

* distinguish editable requirement

This would allow non-editable installs on computers not connected to the internet, which I believe is not uncommon.

* re-introduce pip general requirement

pyproject.toml requires pip>=19.0 to allow the backend selection

* make optional dependencies user lower case to work

Co-authored-by: Andy Thomas <andythomas@web.de>