style(control_dummy): better variable names (snake_case) and correct run_delay use
run_delay should be used to avoid to frequent addressing of the hardware. This was previously done by harming the interactivity upon the start of the program where a wait was inserted at the beginning. Now the wait is done at the end of the loop where it seems to have the same function. Likely more control GUIs should be converted to this way.