fix(visadevice): print out source of the error inside visadevice (#260)
* fix(visadevice): print out source of the error inside visadevice

this should make it easier to identify the source of communication problems upon exceptions

* refactor: remove print statement which no longer is needed

identifying a failing device should now be possible by the message printed during an exception. The only unfortunate thing is that the name of a device inside the system (which was printed previously) is now not known during the error. If we want to also keep it to be printed we would have to catch the exception and print it in a similar way as its done inside VisaDevice

* pep8formatting action fixes (#261)

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>

* fix: print identifier on the system level in case of exceptions during hardware access

* fix(System): better error information in set_/read_/trigger_value

unify error handling in System. More information about the column/device we try to access.

* pep8formatting action fixes

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>