fix(icons): Make icons work in Linux/Gnome (#299)
* fix(icons): install icons for ma7 files correctly in linux/Gnome

For gnome to use the icons with the default icon theme it is necessary to install the mimetype icon to the default icon them. hicolor is not used anymore as a fallback by default

* ci(icons): try to enable manual running of the icon refresh github action

this can be convenient when the automatic mechanism fails to trigger. in the current case it would help to test the github action as well as to make the png icons with a defined size

* ci(icons): fix previous commit

* pep8formatting action fixes (#300)

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>