fix(ControlGUI): avoid triggering value casting
all old control guis might have problems to run with recent controlwindow. The problem is that the value casting can fail when the value is None and the dtype is int...
not sure how this could not be discovered earlier. But it showed up as a clear error at TUD/strain-pi control_blue