make modular control GUI work (#407)
* make modular control GUI work

Ctrl+S can be used to save the active arrangement of windows and their size (tested on windows only).

Current status of the status output (collapsed or open) is also stored by Ctrl+S.

GuiDicts can now be activated and deactivated and all of the things continue to work as one would expect (hopefully, tested to the best of my ability).

Control-dummy has now working commands which can be used in the system integration. (They got broken at some point during the rewrite).

* allow panic mode to be activated from refresh

So far the panic button was the only way to activate the panic mode. Its
not possible to set/unset the panic mode also from the code.

* feat(system_dummygui): add readout for pressure parameter

* remove 'movable' feature from last docked guidict

* simplify System.reset implementation for VisaDevice

* pep8formatting action fixes

* refactor: split set_cmd_funcs ino subfunctions

* allow docking only in the top area

* output helpful info if GuiDict is forced to be docked

---------

Co-authored-by: pheowl <github@rschlitz.de>
Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>