feat(INSTALL): improved install dialog with support for virtual environments (#435)
* feat(INSTALL): improved install dialog

allow creation of virtual enviroments for matr1x.
still needs some more work (assess default options)

* remove --user option, improve python detection

make use of system-site-packages and clear the venv upon repeated creation.

* Fixes by autopep8 action (#436)

* pep8formatting action fixes

* Update install_dialog.py

---------

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>
Co-authored-by: Dominik Kriegner <dominik.kriegner@gmail.com>

* raise required Python version to 3.8 in pyproject.toml

this makes the required Python version equal in the readme and in the
pyproject.toml file. anything older than 3.8 is anyway end of life

* make INSTALL and INSTALL.ps1 line 5 be equal

in this line the used python interpreter can be set by the user. this
should be near the top of the file

* update install section of readme and cleanup

I removed content of the readme which seems to have nothing to do with
ifw software.

* fix(INSTALL): use correct python for venv

* fix(INSTALL): regex requires raw string in 3.12

* enable virtual environment creation in powershell installer

* improve formatting of README.md

* make tests run if matrix is not in the PATH

* add h5py option to windows install script

for some reason this functionality was missing so far

* pep8formatting action fixes (#442)

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>

* Add note about not working python 3.12 to README.md

* enforce changing of the userInstallation option

If it was set previously it would otherwise stay set which is not what
we want. It is still possible to overide this via command line options
to the install GUI or when editing the install.cfg and using a
"headless" installation.

* add headless installation section to the readme

* Update MacOS install in readme

* Move MacOS section up, python -> Python

* update script2bundle

prepare to move to pypi, unify versions in utils and script2bundle repo

* pep8formatting action fixes

* print warning if userInstallation option is set

* print out warning about deleting venv

* make default config options adjusted for an end user

* add note about updating hinting on venvs

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>
Co-authored-by: Andy Thomas <andythomas@web.de>
Co-authored-by: andythomas <andythomas@users.noreply.github.com>