Issue453 (#457)
* feat(INSTALL): add *.matrix file to different INSTALL dialogs

* feat(matrix_script): adds possibility to load matrix-script with an initial script passed as command line parameter

* fix(desktop-integration): also link ma6 files to matrix-preview

* feat(matrix-script): prepare Mac Desktop integration

* pep8formatting action fixes

* feat(matrix-script): link *.matrix on a Mac

* avoid repeated elevated commands, fix icon

every call to invoke-elevated requests the user to allow this. therefore we should keep only one such call

* Its 2023! use utf-8

On windows UTF-8 is surprisingly still not used by all of python. (See PEP686 which should change this in 3.15). This commit enforces our datafile to be utf-8 and also sets an environment variable to start python in a special utf-8 mode. I hope this has no side affects! It solves my angstroem issue.

* (mac) Use proper default location

* (mac) delete app bundles during uninstall

* (Mac): open ma6 with matrix-preview

* pep8formatting action fixes

* UNINSTALL.ps1: merge elevated code blocks to avoid multiple user confirmations

* rename ma7 to datafile

* fix UNINSTALL.ps1

---------

Co-authored-by: Andy Thomas <andythomas@web.de>
Co-authored-by: andythomas <andythomas@users.noreply.github.com>
Co-authored-by: Dominik Kriegner <dominik.kriegner@gmail.com>