AJA-growth system: new temperature controller and fixes (#612)
One noteworthy fix contained here is the change of the default log file filename in control-guis on windows. Windows does not like ":" in filenames. Our default timestamp format therefore caused troubles. I now remove the ":" on windows in the logfile-name. We could also discuss the default time stamp format, but since its usually used just in printouts I think its ok to keep it as is.

The rest of the changes are purely for the AJA system and should be moved to an dedicated repository in the future.

* initial (untested) implementation for the Tempco 9300 PID controller

* add testing control GUI for Tempco9300.

this will be included in control_aja soon.

* fix(control-aja): manually set baud rate in devices

This became necessary after an firmware update of the perle which actually fixed some problems with RS485 devices, but now makes it required to set the baud rate from the device driver. (I believe it was supposed to work also before but somehow did not)

* use separate commands for setpoint and temperature reading in control_tempco

* add new dependency for tempco driver

* fix tempco9300 and update dependencies

* add tempco GuiDict to control-aja

* fix tempco PID property

* add git device to system_aja

* add power ramp possibility by system method

* clean up aja_tasks

* fix new logging filename on windows

":" seems to be a reserved thing for files on windows...

* some printout during power ramp

* clean up control aja

no idea why this code seems to be not needed anymore and actually partly problematic if inside deactivatable GuIDicts

* pep8formatting action fixes (#614)

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>