fix(matrix-preview): allow showing columns with some None values (#652)
If a part of a GuiDict is off but its values are logged we typically get 'None' data values. This is to allow a distinction of NaN and None values but causes problems with the visualization of such data by matrix-preview. This commit adds an option to loadmatrix to allow to automatically replace None values by NaN which makes them viewable. The default behavior of loadmatrix is not changed.

Additionally this commit fixes some linter warnings in the eval.py file.
(code refactoring)