fix(loadmatrix): replace None by 0 instead of NaN (#666)
* fix(loadmatrix): replace None by 0 instead of NaN

loadmatrix failed if an integer column contained a mix of integers and
'None'. loadmatrix also misbehaved for boolean columns where a 'None'
evaluated to True. Now loadmatrix replaces the 'None' values by 0 which
evaluates to False for boolean columns and shows a zero for numerical
columns. This is not the ideal behavior for all column types but
fixes #662 at least temporarily.

* fix(loadmatrix): replace numpy.genfromtxt by pandas.read_csv

* pep8formatting action fixes (#667)

Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>

* fix(loadmatrix): ignore comment lines starting with '#'

* remove 1 point hack which is not needed for pandas

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: dkriegner <dkriegner@users.noreply.github.com>