feat(matrix): add comments to datafile upon aborting (#694)
* feat(matrix): add comments to datafile upon aborting

when a measurement is paused or aborted we did not note this in the datafile. this commit adds a comment functionality which should allow to keep better track of what happened during a measurement. It is a small change to the datafile format which should be testing using the different means of parsing the file. With eval.loadmatrix it works without further changes.

* add user input query upon abort

the user is asked after aborting a measurement if the file shall be
labelled unsuccessful. currently this means that a file comment is added
stating this and that the reset function gets the extra argument
status="unsuccessful"

* fix default filename when using system file outside a python package

* fix matrix-script user interface and double comment on pause

* docs: more detailed comments regarding the decorator use

* fix(matrix-script): restore functionality of code in this PR

* fix(PR-check): avoid formatting changes in main

* fix(PR-check): correctly reference main

* refactor(PR-check): run darker against main

* fix(PR-check): correct reference

* ruff action fixes (#712)

Co-authored-by: andythomas <995840+andythomas@users.noreply.github.com>

---------

Co-authored-by: Andy Thomas <andythomas@web.de>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: andythomas <995840+andythomas@users.noreply.github.com>

BREAKING CHANGE: The datafiles now may contain lines starting with '#' at random positions. This might break some data file parsers.