feat(config): change config to toml format and merge with install.cfg (#759)
* refactor(config): provide converter for ini config file to toml file

* use new toml config in matr1x

* update README with new config file documentation

* fix use of new config, shorten some code lines

* move to using tomli-w for writing toml

* make docstrings more extensive and compatible with numpydoc

* integrate install.cfg in new toml config

* make toml stuff work for Python < 3.11

* make the converter work without requiring extra packages

for this I added two submodules of python packages without dependencies.

* move install dialog/script to use toml config

* add example use of config in system_dummy_feature

we might want to consider to put the config loading directly into the
System code. But I am not sure where yet. the new matr1x.get_config_dict
function can be used by all code which needs to access a subsection of
the config.

If the config loading is central in system one might add the hdf5 option
to be configurable for all systems via the config.

* add config options to visadevice

* remove createconfig option

the only relevant config setting for ifwlib is now set unconditionally
by the installer. this seems possible since other config settings are
now not overwritten/removed anymore.

* fix linter errors in config_functions

* perform tests only in core_libaray

otherwise pytest autodiscovers also the tests for tomli which we do not
need to run in our repo

* add upgrade guide from version 7 to 8.

* add recommended python distribution to overview table

* use https url for submodules

* split comment and dc entry after set

@pheowl please check this change since I did not understand the previous
format and why this was added to "source"

* fix typo

* maybe make darker happy ?

* allow generating a key in a non-existing section

* Fixes by PR-check action (#766)

* ruff action fixes

* Update core_library/matr1x/__init__.py

---------

Co-authored-by: dkriegner <9802901+dkriegner@users.noreply.github.com>
Co-authored-by: Dominik Kriegner <dominik.kriegner@gmail.com>

* make ini_to_toml converter more reliable

* remove install.cfg and instruct user to remove matr1x.conf

* ruff action fixes (#767)

Co-authored-by: dkriegner <9802901+dkriegner@users.noreply.github.com>

* remove bashautocomplete option

Somebody interested in maintaining this would need to overhaul it to
bring it back

closes #758

* ruff action fixes (#769)

Co-authored-by: dkriegner <9802901+dkriegner@users.noreply.github.com>

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: dkriegner <9802901+dkriegner@users.noreply.github.com>