fix(matrix-script): correct handling of carriage return (#1618)
* fix(matrix-script): correct handling of carriage return

PR #1610 introduced a simplification which breaks the update print of
the wait function in matrix script. this again restores this
functionality

* use appendPlainText for normal operation

avoid touching the cursor of the user. this now allows to select text
while new output is produced.

* add test method to keep carriage return behavior working

We broke this by regressions already few times. hopefully this helps.

* fix weird behavior identified by test code

previously not the full line was removed in the test. the behavior was
correct in the GUI. Now these two align. This makes the GUI code more
robust against timing changes.

* fix new test on windows.

increased timeout is needed. debugging of such problems is difficult since matrix-script swallows all print output. this means a debug print in the test does not produce visible output.